incorporate review feedback

+++ src/include/miopen/conv/kernel_interface/winograd_kernel_interface.hpp
@@ -28,13 +28,14 @@
 
 #include <type_traits>
 #include <ostream>
+#include <stdint.h>
 
 namespace miopen {
 namespace conv {
 
 struct ProblemDescription;
 
-enum class WinoShaderFlagsV2 : unsigned long long
+enum class WinoShaderFlagsV2 : uint64_t
 {
     F_REVERSE_R                  = 1ULL << 0,
     F_REVERSE_S                  = 1ULL << 1,
@@ -90,38 +91,34 @@ inline std::ostream& operator<<(std::ostream& s, const WinoShaderActivationModeV
 struct WinoShaderArgsV2
 {
     // Main convolution parameters
-    unsigned int N;     // batch size
-    unsigned int C;     // number of input channels in each filter group
-    unsigned int H;     // input height
-    unsigned int W;     // input width
-    unsigned int K;     // number of output channels in each filter group
-    unsigned int R;     // filter height
-    unsigned int S;     // filter width
-    int32_t pad_h;      // padding in h dimension
-    int32_t pad_w;      // padding in w dimension
-    unsigned int out_h; // output height
-    unsigned int out_w; // output width
-    unsigned int G;     // number of filter groups
+    uint32_t N;     // batch size
+    uint32_t C;     // number of input channels in each filter group
+    uint32_t H;     // input height
+    uint32_t W;     // input width
+    uint32_t K;     // number of output channels in each filter group
+    uint32_t R;     // filter height
+    uint32_t S;     // filter width
+    int32_t pad_h;  // padding in h dimension
+    int32_t pad_w;  // padding in w dimension
+    uint32_t out_h; // output height
+    uint32_t out_w; // output width
+    uint32_t G;     // number of filter groups
 
     // Data layout related parameters
-    unsigned int
-        d_N_stride; // stride in number of elements of the N dimension of the input data buffer
-    unsigned int
-        d_C_stride; // stride in number of elements of the C dimension of the input data buffer
-    unsigned int
-        d_H_stride; // stride in number of elements of the H dimension of the input data buffer
-    unsigned int
-        d_G_stride; // stride in number of elements of the G dimension of the input data buffer
+    uint32_t d_N_stride; // stride in number of elements of the N dimension of the input data buffer
+    uint32_t d_C_stride; // stride in number of elements of the C dimension of the input data buffer
+    uint32_t d_H_stride; // stride in number of elements of the H dimension of the input data buffer
+    uint32_t d_G_stride; // stride in number of elements of the G dimension of the input data buffer
 
-    unsigned int f_K_stride; // stride in number of elements of the K dimension of the filter buffer
-    unsigned int f_C_stride; // stride in number of elements of the C dimension of the filter buffer
-    unsigned int f_R_stride; // stride in number of elements of the R dimension of the filter buffer
-    unsigned int f_G_stride; // stride in number of elements of the G dimension of the filter buffer
+    uint32_t f_K_stride; // stride in number of elements of the K dimension of the filter buffer
+    uint32_t f_C_stride; // stride in number of elements of the C dimension of the filter buffer
+    uint32_t f_R_stride; // stride in number of elements of the R dimension of the filter buffer
+    uint32_t f_G_stride; // stride in number of elements of the G dimension of the filter buffer
 
-    unsigned int o_N_stride; // stride in number of elements of the N dimension of the output buffer
-    unsigned int o_K_stride; // stride in number of elements of the K dimension of the output buffer
-    unsigned int o_H_stride; // stride in number of elements of the H dimension of the output buffer
-    unsigned int o_G_stride; // stride in number of elements of the G dimension of the output buffer
+    uint32_t o_N_stride; // stride in number of elements of the N dimension of the output buffer
+    uint32_t o_K_stride; // stride in number of elements of the K dimension of the output buffer
+    uint32_t o_H_stride; // stride in number of elements of the H dimension of the output buffer
+    uint32_t o_G_stride; // stride in number of elements of the G dimension of the output buffer
 
     // Fused activation parameters
     float alpha;                                  // activation parameter alpha
@@ -129,7 +126,7 @@ struct WinoShaderArgsV2
     WinoShaderActivationModeV2_t activation_mode; // activation mode
 
     // Other shader parameters
-    unsigned int n_groups;     // number of shader groups
+    uint32_t n_groups;         // number of shader groups
     WinoShaderFlagsV2 flags64; // shader flags
     uint8_t sync_limit;        // maximum number of sync attempts
     uint8_t sync_period;       // synchronization period
@@ -137,7 +134,7 @@ struct WinoShaderArgsV2
     bool SetConvParams(const ProblemDescription& problem);
     void SetStrides(const ProblemDescription& problem);
     void SetActivParams(WinoShaderActivationModeV2_t mode, float alpha, float beta) noexcept;
-    void SetShaderParams(unsigned int n_groups,
+    void SetShaderParams(uint32_t n_groups,
                          WinoShaderFlagsV2 flags,
                          uint8_t sync_limit,
                          uint8_t sync_period) noexcept;
